pyproject.toml
[tool.poetry]
name = "Formidable backend"
version = "0.1.0"
description = "A form management system API"
authors = ["Dan Percic <danpercic86@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
psycopg2-binary = "^2.9.1"
djangorestframework = "^3.12.4"
django-cors-headers = "^3.8.0"
Pillow = "^8.3.1"
django-admin-thumbnails = "^0.2.6"
PyYAML = "^5.4.1"
docutils = "^0.17.1"
django-model-utils = "^4.1.1"
drf-spectacular = "^0.18.0"
django-jazzmin = "^2.4.8"
django-tabular-permissions = "^2.8"
dj-rest-auth = { extras = ["with_social"], version = "^2.1.11" }
djangorestframework-simplejwt = "^4.8.0"
drf-extensions = "^0.7.1"
django-admin-sortable2 = "^1.0"
django-ckeditor = "^6.1.0"

[tool.poetry.dev-dependencies]
pylint = "^2.10.2"
black = "^21.7b0"
pycodestyle = "^2.7.0"
django-querycount = "^0.7.0"
django-debug-toolbar = "^3.2.2"
nplusone = "^1.0.0"
django-extra-checks = "^0.11.0"
factory-boy = "^3.2.0"
coverage = { extras = ["toml"], version = "5.5" }
mypy = "^0.910"
django-extensions = "^3.1.3"

[tool.black]
target-version = ['py38']
exclude = '''
/(
    \.git
  | \__pycache__
  | \venv
  | \.venv
)/
'''

[tool.pylint.messages_control]
disable = """
    C0330,
    C0326,
    missing-class-docstring,
    missing-module-docstring,
    too-few-public-methods,
    unused-wildcard-import,
    wildcard-import,
    unsubscriptable-object,
    missing-function-docstring,
    ungrouped-imports,
    no-member,
    too-many-ancestors,
    import-outside-toplevel,
    protected-access,
    invalid-name
"""

[tool.pylint.master]
ignore = """
    migrations
"""

[tool.pylint.format]
max-line-length = "88"

[tool.coverage.run]
omit = [".venv/*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 100.0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
